feat: add TypeScript setup and split frontend
