+++ Makefile
@@ -1,0 +1,3 @@
+.PHONY: build-frontend
+build-frontend:
+	npm run build
