+++ package.json
@@ -1,0 +1,11 @@
+{
+  "name": "docrouter",
+  "version": "1.0.0",
+  "type": "module",
+  "devDependencies": {
+    "typescript": "^5.6.3"
+  },
+  "scripts": {
+    "build": "tsc"
+  }
+}
