+++ src/web_app/static/main.ts
@@ -1,0 +1,10 @@
+import { setupUpload } from './upload';
+import { setupFolders, refreshFolderTree } from './folders';
+import { setupChat } from './chat';
+
+document.addEventListener('DOMContentLoaded', () => {
+  setupUpload();
+  setupFolders();
+  setupChat();
+  refreshFolderTree();
+});
