+++ src/web_app/templates/index.html
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="/static/style.css">
     <link href="https://unpkg.com/cropperjs@1.5.13/dist/cropper.min.css" rel="stylesheet" />
     <script src="https://unpkg.com/cropperjs@1.5.13/dist/cropper.min.js" defer></script>
-    <script src="/static/main.js" defer></script>
+    <script type="module" src="/static/dist/main.js"></script>
 </head>
 <body>
     <div class="container">
