+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "target": "ES6",
+    "module": "ES6",
+    "rootDir": "src/web_app/static",
+    "outDir": "src/web_app/static/dist",
+    "strict": false,
+    "esModuleInterop": true
+  },
+  "include": ["src/web_app/static/**/*.ts"],
+  "exclude": ["src/web_app/static/dist"]
+}
